ml-pipeline/requirements.txt
# Wildlife MLOps Platform - Python Dependencies

# Core ML
torch>=2.0
torchvision>=0.15
scikit-learn>=1.3

# MLOps
mlflow>=2.10
dvc>=3.0

# Data
pandas>=2.0
numpy>=1.24
Pillow>=10.0

# Inference API
fastapi>=0.110
uvicorn>=0.27
python-multipart>=0.0.6

# Monitoring
evidently>=0.4
